Add DeveloperKey feature for copying NBT data
Introduced the DeveloperKey feature, allowing developers to copy NBT data of entities, blocks, and items to the clipboard using a keybind. Added a HandledScreenAccessor mixin to access the focused slot in handled screens. Updated FeatureUtil to register the new feature and modified nar.mixins.json to include the new accessor. Also added utility methods in NotARat for sending chat messages.